requirements.txt
streamlit>=1.28.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
vtracer>=0.6.0
rembg>=2.0.0
onnxruntime>=1.15.0
